refactor: extract quickpay ledger and spend store
